+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added support for parallel deployment. The `deploy` command will now deploy modules in parallel by default.
+- Added `--threads`/`-t` to control how many threads to use to deploy in parallel. Also, added `--serial`/`-k` to force serial deployment.
 
 ## [0.7.4-beta] - 2021-04-10
 ### Added